Wednesday and Tuesday
Worked on getting closer to having events do more than display text. Involved designing a way to format text files so that the code can parse them and generate events and GameObjects (a new subclass, parent to interactable, and by extension player) and place them into dictionaries for easy lookup and access. At this point, the events are not being initialized properly. Events are generated with a name, but dialogue is not showing up. Some testing and examining of the loops that generate events are required.